simulator: move Bandwidth to dedicated module
